Renamed quantities module to more appropriate eproperties
