+++ skued/tests/test_eproperties.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 import numpy as np
 import unittest
-from ..quantities import lorentz, electron_wavelength, interaction_parameter
+from .. import lorentz, electron_wavelength, interaction_parameter
 
 class TestLorentz(unittest.TestCase):
